+++ src/main/java/de/jeff_media/BestTools/BlockPlaceListener.java
@@ -28,7 +28,7 @@ public class BlockPlaceListener implements Listener {
 
         if(item.getAmount() != 1) return;
 
-        if(!p.hasPermission("besttool.refill")) return;
+        if(!p.hasPermission("besttools.refill")) return;
         if(!playerSetting.refillEnabled) {
             if(!playerSetting.hasSeenRefillMessage) {
                 p.sendMessage(main.messages.MSG_REFILL_USAGE);