+++ src/main/java/de/jeff_media/BestTools/RefillListener.java
@@ -5,32 +5,45 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockPlaceEvent;
+import org.bukkit.event.player.PlayerItemConsumeEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.PlayerInventory;
 
-public class BlockPlaceListener implements Listener {
+public class RefillListener implements Listener {
 
     Main main;
 
-    BlockPlaceListener(Main main) {
+    RefillListener(Main main) {
         this.main=main;
     }
 
     @EventHandler
-    public void onPlace(BlockPlaceEvent event) {
-        main.debug("BlockPlace");
+    public void onSnackDigestion(PlayerItemConsumeEvent event) {
         Player p = event.getPlayer();
+        attemptRefill(p);
+    }
+
+    @EventHandler
+    public void onPlace(BlockPlaceEvent event) {
+        Player p = event.getPlayer();
+        attemptRefill(p);
+    }
+
+    private void attemptRefill(Player p) {
+        if(!PlayerUtils.isAllowedGamemode(p,main.getConfig().getBoolean("allow-in-adventure-mode"))) {
+            return;
+        }
         PlayerInventory inv = p.getInventory();
         PlayerSetting playerSetting = main.getPlayerSetting(p);
         ItemStack item = inv.getItemInMainHand();
         Material mat = item.getType();
         int currentSlot = inv.getHeldItemSlot();
 
-        if(item.getAmount() != 1) return;
+        if (item.getAmount() != 1) return;
 
-        if(!p.hasPermission("besttool.refill")) return;
-        if(!playerSetting.refillEnabled) {
-            if(!playerSetting.hasSeenRefillMessage) {
+        if (!p.hasPermission("besttool.refill")) return;
+        if (!playerSetting.refillEnabled) {
+            if (!playerSetting.hasSeenRefillMessage) {
                 p.sendMessage(main.messages.MSG_REFILL_USAGE);
                 playerSetting.setHasSeenRefillMessage(true);
             }
@@ -38,9 +51,9 @@ public class BlockPlaceListener implements Listener {
             return;
         }
 
-        int refillSlot = RefillUtils.getMatchingStackPosition(inv,mat,inv.getHeldItemSlot());
-        if(refillSlot != -1) {
-            main.refillUtils.refillStack(inv,refillSlot,currentSlot,inv.getItem(refillSlot));
+        int refillSlot = RefillUtils.getMatchingStackPosition(inv, mat, inv.getHeldItemSlot());
+        if (refillSlot != -1) {
+            main.refillUtils.refillStack(inv, refillSlot, currentSlot, inv.getItem(refillSlot));
         }
     }
 }