+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 1.0.1
+- Fixed typo in permission. /refill should work now
 - Added Spanish translation
 - Added download link to Update checker