+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+- Refill works on food, snacks, and anything else digestible
+- /besttools and /refill will only work in survival mode or, if enabled in the config, in adventure mode
+- Added German translation 
+
 ## 1.0.1
 - Added Spanish translation
 - Added download link to Update checker